connections: Add user information to connections based on authToken
This is done via a new UserFromAuthToken function that's passed to
the connection as part of its (also new) connection config.

When the auth token is received from the GQL_CONNECTION_INIT message,
this hook is called and the returned user (can be anything) is stored
in the connection. It can later be retrieved from the connection
via the User() method.