subscriptions: Convert GraphQL errors to nil if they are empty
This reduces the likelyhood of servers that use graphqlws to send
an empty array of errors along with subscription updates to clients.

This also makes subscriptions work better in GraphiQL, where
previously, the empty array of errors would make
GraphiQL-Subscriptions-Fetcher would treat all updates as errors:

https://github.com/apollographql/GraphiQL-Subscriptions-Fetcher/blob/master/src/fetcher.ts#L36

(An empty array is considered truthy).